+++ web-server/src/main/java/com/se14/service/ProjectService.java
@@ -14,6 +14,6 @@ public interface ProjectService {
     Map<String, Object> getStatistic(Project project);
     List<User> listUser(Project project, UserRole role);
     boolean hasUser(Project project,User user);
-    Project findProjectById(long id);
+    Project findProjectById(Integer id);
     List<Project> findProjectByUser(User user);  // 현석님 요청. 새로운 메서드 추가
 }
